file fix

+++ solar_input.py
@@ -56,19 +56,20 @@ def parse_star_parameters(line, star):
 
     splitted = line.split()
 
-    star.m = int(splitted[3])
+    star.m = float(splitted[3])
 
-    star.x = int(splitted[4])
+    star.x = float(splitted[4])
 
-    star.y = int(splitted[5])
+    star.y = float(splitted[5])
 
-    star.Vx = int(splitted[6])
+    star.Vx = float(splitted[6])
 
-    star.Vy = int(splitted[7])
+    star.Vy = float(splitted[7])
 
-    star.r = int(splitted[1])
+    star.r = float(splitted[1])
 
     star.color = splitted[2]
+
 
 def parse_planet_parameters(line, planet):
     """Считывает данные о планете из строки.
@@ -90,17 +91,17 @@ def parse_planet_parameters(line, planet):
 
     splitted = line.split()
 
-    planet.m = int(splitted[3])
+    planet.m = float(splitted[3])
 
-    planet.x = int(splitted[4])
+    planet.x = float(splitted[4])
 
-    planet.y = int(splitted[5])
+    planet.y = float(splitted[5])
 
-    planet.Vx = int(splitted[6])
+    planet.Vx = float(splitted[6])
 
-    planet.Vy = int(splitted[7])
+    planet.Vy = float(splitted[7])
 
-    planet.r = int(splitted[1])
+    planet.r = float(splitted[1])
 
     planet.color = splitted[2]
